+++ apps/ui/src/app/routes/apps/deploymentDetail.tsx
@@ -2,7 +2,7 @@ import { FC, useEffect, useMemo, useState, useRef } from 'react';
 import { useParams } from 'react-router-dom';
 import * as Tabs from '@radix-ui/react-tabs';
 import * as Tooltip from '@radix-ui/react-tooltip';
-import { UilExclamationTriangle, UilLock, UilLockSlash, UilSpinner } from '@iconscout/react-unicons';
+import { UilExclamationTriangle, UilExternalLinkAlt, UilLock, UilLockSlash, UilSpinner } from '@iconscout/react-unicons';
 import { Utils } from '@secretarium/crypto';
 import * as NivoGeo from '@nivo/geo';
 import Ansi from 'ansi-to-react';
@@ -24,9 +24,32 @@ export const AppDeploymentDetail: FC = () => {
     const [shouldAutoScroll] = useState(false);
     const [effectiveClusterFQDN, setEffectiveClusterFQDN] = useState<string>();
     const [WASMFingerprint, setWASMFingerprint] = useState<string>();
+    const { data: secretariumNodeInfo, isLoading: isLoadingSecretariumNodeInfo } = api.v0.system.getSecretariumNode.useQuery();
+    const { data: uiHostingInfo } = api.v0.system.getUIHostingDomain.useQuery();
     const { data: deployment, isLoading: isLoadingDeployments } = api.v0.deployments.getById.useQuery({ deploymentId: deploymentId ?? '' }, {
         refetchInterval: (s) => ['errored', 'terminated', 'deployed'].includes(s.state.data?.status ?? '') ? (Date.now() - (s.state.data?.createdAt.getTime() ?? 0) < 60000 ? 5000 : 60000) : 500
     });
+
+    const uiHostingDomain = useMemo(() => {
+        if (!uiHostingInfo)
+            return undefined;
+        const url = URL.parse(uiHostingInfo);
+        if (url)
+            return url.host;
+        return undefined;
+    }, [uiHostingInfo]);
+
+    const secretariumNode = useMemo(() => {
+        if (!secretariumNodeInfo)
+            return undefined;
+        const node = secretariumNodeInfo.split('|')[0]
+        if (!node)
+            return undefined;
+        const url = URL.parse(node);
+        if (url)
+            return url.host;
+        return undefined;
+    }, [secretariumNodeInfo]);
 
     useEffect(() => {
 
@@ -118,7 +141,7 @@ export const AppDeploymentDetail: FC = () => {
         };
     }), [geoFeatures.features, datacentreMarkers]);
 
-    if (isLoadingDeployments || !deployment)
+    if (isLoadingDeployments || !deployment || isLoadingSecretariumNodeInfo)
         return <>
             We are fetching data about your deployment.<br />
             It will only take a moment...<br />
@@ -136,7 +159,7 @@ export const AppDeploymentDetail: FC = () => {
         }).catch(() => { return; });
     }
 
-    if (!deploymentAddress)
+    if (!deploymentAddress || !secretariumNode)
         return <>
             We experienced an issue fetching your deployment.<br />
         </>;
@@ -265,6 +288,7 @@ export const AppDeploymentDetail: FC = () => {
             ? <Tabs.Root defaultValue="inspect" className='flex flex-col w-full'>
                 <Tabs.List className='flex flex-shrink-0 border-b'>
                     <Tabs.Trigger value="inspect" className='flex-1 border-0 border-b-2 border-transparent rounded-none data-[state=active]:border-klave-light-blue shadow-none text-sm font-normal text-gray-600 dark:text-gray-400 hover:text-klave-light-blue data-[state=active]:font-bold data-[state=active]:text-klave-light-blue'>Inspect</Tabs.Trigger>
+                    {deployment.buildOutputHasUI ? <Tabs.Trigger value="ui" className='flex-1 border-0 border-b-2 border-transparent rounded-none data-[state=active]:border-klave-light-blue shadow-none text-sm font-normal text-gray-600 dark:text-gray-400 hover:text-klave-light-blue data-[state=active]:font-bold data-[state=active]:text-klave-light-blue'>Embedded UI</Tabs.Trigger> : null}
                     <Tabs.Trigger value="configuration" className='flex-1 border-0 border-b-2 border-transparent rounded-none data-[state=active]:border-klave-light-blue shadow-none text-sm font-normal text-gray-600 dark:text-gray-400 hover:text-klave-light-blue data-[state=active]:font-bold data-[state=active]:text-klave-light-blue'>Configuration</Tabs.Trigger>
                     <Tabs.Trigger value="build" className='flex-1 border-0 border-b-2 border-transparent rounded-none data-[state=active]:border-klave-light-blue shadow-none text-sm font-normal text-gray-600 dark:text-gray-400 hover:text-klave-light-blue data-[state=active]:font-bold data-[state=active]:text-klave-light-blue'>Build and Dependencies</Tabs.Trigger>
                     <Tabs.Trigger value="attest" className='flex-1 border-0 border-b-2 border-transparent rounded-none data-[state=active]:border-klave-light-blue shadow-none text-sm font-normal text-gray-600 dark:text-gray-400 hover:text-klave-light-blue data-[state=active]:font-bold data-[state=active]:text-klave-light-blue'>Attest</Tabs.Trigger>
@@ -294,12 +318,10 @@ export const AppDeploymentDetail: FC = () => {
                         </pre>
                     </div>
                 </Tabs.Content>
-                <Tabs.Content value="configuration">
+                <Tabs.Content value="ui">
                     <div className='mt-10'>
-                        <h2 className='font-bold mb-3'>Build configuration for this deployment</h2>
-                        <pre className='overflow-auto whitespace-pre-wrap break-words w-full max-w-full bg-gray-100 p-3'>
-                            {JSON.stringify(deployment.configSnapshot ?? '', null, 4)}
-                        </pre>
+                        <h2 className='font-bold mb-3'>UI configured to deploy with this application <a title="Open in a new tab or window" href={`https://${deployment.id}.${uiHostingDomain}?g=${secretariumNode}&d=${fqdn}`} target='_blank' rel='noreferrer' className='text-klave-light-blue hover:underline inline-block'><UilExternalLinkAlt className='inline-block h-4' /></a></h2>
+                        <iframe className='w-full h-[80vh] border-0' src={`https://${deployment.id}.${uiHostingDomain}?g=${secretariumNode}&d=${fqdn}`} title={`UI for ${fqdn}`} sandbox="allow-same-origin allow-scripts  allow-modals allow-forms allow-popups allow-presentation" />
                     </div>
                 </Tabs.Content>
                 <Tabs.Content value="build">