+++ libs/providers/src/index.ts
@@ -1,6 +1,7 @@
 export * from './lib/probot';
 export * from './lib/logger';
 export * from './lib/secretarium';
+export * from './lib/objectStore';
 export * from './lib/githubReconciler';
 export * from './lib/environmentPopulator';
 export * from './lib/dispatcher';