+++ libs/api/CHANGELOG.md
@@ -1,4 +1,22 @@
 # Changelog
+## [1.4.1](https://github.com/klave-network/platform/compare/api@1.4.0...api@1.4.1) (2025-05-08)
+
+### Dependency Updates
+
+* `ui-kit` updated to version `1.4.0`
+## [1.4.0](https://github.com/klave-network/platform/compare/api@1.3.4...api@1.4.0) (2025-05-08)
+
+### Dependency Updates
+
+* `db` updated to version `1.3.4`
+* `providers` updated to version `1.3.4`
+* `constants` updated to version `1.3.4`
+* `ui-kit` updated to version `1.3.4`
+
+### Features
+
+* **db,api,providers,hubber,ui,dispatcher,constants:** Add Embedded UI Hosting ([5a748c8](https://github.com/klave-network/platform/commit/5a748c8b616ad3674058517f5bd9c040361fc3c4))
+
 ## [1.3.4](https://github.com/klave-network/platform/compare/api@1.3.3...api@1.3.4) (2025-04-30)
 
 ### Dependency Updates