+++ libs/db/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## [1.4.0](https://github.com/klave-network/platform/compare/db@1.3.2...db@1.4.0) (2025-05-08)
+
+
+### Features
+
+* **db,api,providers,hubber,ui,dispatcher,constants:** Add Embedded UI Hosting ([5a748c8](https://github.com/klave-network/platform/commit/5a748c8b616ad3674058517f5bd9c040361fc3c4))
+
 ## [1.3.2](https://github.com/klave-network/platform/compare/db@1.3.1...db@1.3.2) (2025-04-30)